% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_renum.R
\name{blup}
\alias{blup}
\title{blup}
\usage{
blup(
  datarenum,
  formula,
  fields_output = NULL,
  weights_object = NULL,
  residual_start = 1,
  VCA_RRM = NULL,
  VCD_RRM = NULL,
  VCA = 0.1,
  VCD = 0.1,
  ped_name,
  PED_DEPTH = 3,
  genotype_file,
  missing_values = 0,
  RRM_option,
  het_res_variance = NULL,
  fit_option,
  run_model = list(method = "blupf90+", dense = F, Gibbs_option = NULL, slurm_option =
    NULL, n_threads = 1),
  cross_validation = list(group_name_cv, LOO_cv, prop, nrep, eff_name, RRM_option,
    keep_folder = F, parameter_file = "renf90.par", logname = "blupf90.log"),
  keep_files = T,
  extra.option.blup = NULL
)
}
\arguments{
\item{datarenum}{The dataset used for the analysis.}

\item{formula}{The formula specifying the model for the analysis.}

\item{fields_output}{(optional) A character vector specifying the fields to be included in the output.}

\item{weights_object}{(optional) The weights object to be used in the analysis.}

\item{residual_start}{The starting value for residual matrix.}

\item{VCA_RRM}{(optional) The genetic variance matrix for analysis using random regression model.}

\item{VCD_RRM}{(optional) The random effect (non-genetic effect) variance matrix for analysis using random regression model.}

\item{VCA}{The genetic variance matrix for analysis.}

\item{VCD}{The random effect (non-genetic effect) variance matrix for analysis.}

\item{ped_name}{The name of the pedigree file.}

\item{PED_DEPTH}{The depth of the pedigree.}

\item{genotype_file}{The genotype file used in the analysis.}

\item{missing_values}{The value representing missing data in the dataset. Default value is -99.}

\item{RRM_option}{The random regression option option for the analysis.}

\item{het_res_variance}{(optional) The heterogeneity of residual variance parameters.}

\item{fit_option}{The fit options for the analysis.}

\item{run_model}{A list specifying the method and options for running the model.}

\item{cross_validation}{A list specifying the cross-validation options.}

\item{keep_files}{(optional) Logical value indicating whether to keep intermediate files generated during the analysis.}

\item{extra.option.blup}{(optional) Additional options specific to the BLUP analysis. See more options in: http://nce.ads.uga.edu/wiki/doku.php?id=readme.blupf90plus}
}
\description{
blup is a standard analytical function in R for single and multi-trait methods and single-step analysis in the breeding pipeline using
BLUPF90 software
}
\details{
The `blup` function performs Best Linear Unbiased Prediction (BLUP) analysis on the given dataset using the specified model formula.
It estimates genetic parameters, such as additive genetic variance and non-genetic random effects, using the provided input parameters.
 Various options are available to customize the analysis, including random regression option using Legendre polynomials,
 residual heritability, pedigree information, genotype data, single-step GBLUP.
}
\examples{
\dontrun{
# Example usage of blup function
blup(datarenum, formula, fields_output = c("field1", "field2"), weights_object = weights,
     residual_start = 1, VCA_RRM = VCA_matrix, VCD_RRM = VCD_matrix, VCA = 0.1, VCD = 0.1,
     ped_name = "pedigree.txt", PED_DEPTH = 3, genotype_file = "genotypes.txt", missing_values = -999,
     RRM_option = "option1", het_res_variance = 0.2, fit_option = "option2",
     run_model = list(method = "blupf90+", dense = FALSE, Gibbs_option = NULL, slurm_option = NULL,
                      n_threads = 1),
     cross_validation = list(group_name_cv, LOO_cv, prop, nrep, eff_name, RRM_option,
                            keep_folder = FALSE, parameter_file = "renf90.par", logname = "blupf90.log"),
     keep_files = TRUE, extra.option.blup = NULL)
}

}
\references{
\itemize{
  \item Lourenco, D., S. Tsuruta, I. Aguilar, Y. Masuda, M. Bermann, A. Legarra, and I. Misztal. "Recent updates in the BLUPF90 software suite." In Proceedings of 12th World Congress on Genetics Applied to Livestock Production (WCGALP) Technical and species orientated innovations in animal breeding, and contribution of genetics to solving societal challenges, pp. 1530-1533. Wageningen Academic Publishers, 2022.

  \item Lourenco, Daniela, Andres Legarra, Shogo Tsuruta, Yutaka Masuda, Ignacio Aguilar, and Ignacy Misztal. "Single-step genomic evaluations from theory to practice: using SNP chips and sequence data in BLUPF90." Genes 11, no. 7 (2020): 790.

}
}
\author{
Leonardo S. Gloria, Luiz F. Brito, Alencar Xavier, Keith A. Cherkauer, Victor B. Pedrosa, Katy Martin Rainey
}
