% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_BLUF90.R
\name{download_BLUPF90}
\alias{download_BLUPF90}
\title{download_BLUPF90}
\usage{
download_BLUPF90(dest_folder = paste0(.libPaths()[1], "/blupf90"))
}
\arguments{
\item{dest_folder}{(optional) A character string specifying the destination folder where the BLUPF90 files will be saved.}
}
\value{
Details:
The \code{download_BLUPF90} function downloads BLUPF90 software files from the appropriate URL based on the operating system. The function identifies the operating system (Linux, Mac_OSX, or Windows) and constructs the URL accordingly. It retrieves the list of available BLUPF90 files from the URL and compares them to the files in the local destination folder. It then downloads the missing files from the URL to the specified destination folder.
}
\description{
Description:
Download all BLUPF90 software from their official repository (http://nce.ads.uga.edu/html/projects/programs/),
and saves them to the destination folder.
}
\details{
Usage:
\code{download_BLUPF90(dest_folder = NULL)}

Arguments:
}
\examples{
\code{
# Download BLUPF90 files to the default destination folder
download_BLUPF90()

# Download BLUPF90 files to a specific destination folder
download_BLUPF90(dest_folder = "~/blupf90_files")
}

Note:
The \code{download_BLUPF90} function requires the \code{RCurl} package to be installed.

Please ensure that you have proper permissions to access the destination folder and that you have an active internet connection during the execution of this function.

}
